temp progress with regards to David's suggestions
