+++ core/src/test/resources/reference.conf
@@ -118,6 +118,7 @@ multiCloudWorkspaces {
     enabled = true
     azureConfig {
         alphaFeatureGroup = "fake_alpha_group"
+
         landingZoneDefinition = "fake_landing_zone_definition"
         landingZoneVersion = "fake_landing_zone_version"
         landingZoneParameters = {
